Remove try-catch and add sample config data
Don't need try-catch to check for config file and provide a fallback -
this is not how config-helper is designed to work. Instead we should
prvide a sample file in the config directory which will be copied to the
local project on `gulp setup`. Then if this doesn't exist, the gulp task
should correctly fail.