Fix semver error with backstop
There has been a major 3.0.0 release of Backstop. Even though we had
`"backstopjs": "^2.6.11"` in the package.json we were still getting the
new changes, specifically
https://github.com/garris/BackstopJS/blob/0535303326e658e69d78f6405aac390b620ff222/capture/genBitmaps.js#L203
which changed vp.name to vp.label which broke everything.

I've changed the package.json to `"backstopjs": "~2.6.11",` in the
meantime and this fixes the errors.