Instantiate a new panini instance in each gulp task and patch panini to refresh the partials in all renderers
Panini exports an singleton, through that it holds all options and they can not be overwritten. This change adds a new panini instance for each task to hold it's own options.

Panini is patched because Handlebars is also a singleton, so the partials will be overwritten. You can find more infos here: https://github.com/zurb/panini/issues/81#issuecomment-285106782